+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog for OnMyShelf web interface
 
 ## 1.0 Beta 2 (202X-XX-XX)
+- Display mode is saved in the local storage
 - Rename 'fields' to 'properties' for better understanding
 - Various bugfixes
+- Upgrade VueJS 3.2.26, VueJS router 4.0.12 and axios 0.24.0
 
 ## 1.0 Beta 1 (2021-10-09)
 - Create, edit & delete collections